+++ src/org/jgroups/protocols/pmcast/DeliverProtocol.java
@@ -1,4 +1,4 @@
-package org.jgroups.groups;
+package org.jgroups.protocols.pmcast;
 
 import org.jgroups.Message;
