+++ src/org/jgroups/protocols/pmcast/GROUP_MULTICAST.java
@@ -7,16 +7,13 @@ import org.jgroups.View;
 import org.jgroups.annotations.MBean;
 import org.jgroups.annotations.ManagedAttribute;
 import org.jgroups.annotations.ManagedOperation;
-import org.jgroups.groups.DeliverProtocol;
-import org.jgroups.groups.GroupAddress;
-import org.jgroups.groups.MessageID;
-import org.jgroups.groups.header.GroupMulticastHeader;
-import org.jgroups.groups.manager.DeliverManagerImpl;
-import org.jgroups.groups.manager.SenderManager;
-import org.jgroups.groups.manager.SequenceNumberManager;
-import org.jgroups.groups.stats.StatsCollector;
-import org.jgroups.groups.threading.DeliverThread;
-import org.jgroups.groups.threading.SenderThread;
+import org.jgroups.protocols.pmcast.header.GroupMulticastHeader;
+import org.jgroups.protocols.pmcast.manager.DeliverManagerImpl;
+import org.jgroups.protocols.pmcast.manager.SenderManager;
+import org.jgroups.protocols.pmcast.manager.SequenceNumberManager;
+import org.jgroups.protocols.pmcast.stats.StatsCollector;
+import org.jgroups.protocols.pmcast.threading.DeliverThread;
+import org.jgroups.protocols.pmcast.threading.SenderThread;
 import org.jgroups.stack.Protocol;
 
 import java.util.Set;
