+++ src/org/jgroups/protocols/pmcast/threading/SenderThread.java
@@ -1,4 +1,4 @@
-package org.jgroups.groups.threading;
+package org.jgroups.protocols.pmcast.threading;
 
 import org.jgroups.Address;
 import org.jgroups.Event;
